Initial commit
SLZ stands for "StateLess Zip". It is intended as a zlib alternative
for HTTP servers and proxies. It aims at compressing small memory
blocks very quickly with minimal CPU overhead and little to no per-
stream memory usage, while keeping compatibility with the gzip format
that all browsers can decompress. It is also possible that it could
make protocols such as SPDY cheaper to implement. The first version
will focus on gzip format for convenience reasons, but if it works
well, a deflate version might be done as well.