MINOR: Initial version
This is a very first experiment. The code is not performant at all,
it relies on eb32tree to quickly look up entries, but probably is
the best we'll get using eb32tree anyway. At least it's enough to
validate several search algorithms before trying to optimize the
code itself.