exp: different model using a simple hash
We now store the last position in a simple hash table. The 4 first
bytes are hashed and give this position which is then checked. The
compression ratio is very good, as 75% of input chars were emitted
as references with a hash table of 16 bits :

    totin=767990000 lit=99110000 ref=576950000

Hash sizes down to 13 bits still give good results. Interestingly,
the results are not as good when indexing only the 3 first chars,
because of the high collision rate which limits the matching length.

The speed as good as well, with 286 MB/s HTML in.