exp: move the hashed pointer to 64 bits and stuff the data in it.
We now store 64 bit values. The lowest part is the data, and the
highest part is the offset. That allows us to avoid a memory access,
because once the hashed entry is retrieved, it is already possible
to check the data.

The only problem is that it requires that the hash table is cleared
upon each access. And intermediary solution could consist in just
using this test as a hint to avoid the memory access as often as
possible.

Note that data are now checked 32 bit at a time. It happens that
not handling 3-byte matches anymore slightly improves the amount of
matching references, which is contrary to expectations.