exp: start to implement a gzip-compatible encoder
This is "rfc1952", the stateless gzip encoder. For now it only supports
literals and copies the input on the output but respects the correct
gzip block format and CRC32. Some tests have been run on a few CRC32
implementations but the observations are that CRC32 will always be
very slow compared to compression and that barring use of arch-specific
asm instructions like PCLMULQDQ, it's very likely that ADLER32 will
have to be implemented and preferred over gzip when possible (zlib
format).