exp: add a configurable limit for the maximum lookup attempts
Here we experiment with the number of entries we need to memorize
in order to find a statistically long hash. It shows that an ideal
non-colliding hash already saves 7.4%, and that looking up a second
location improves this further by 3.6% to achieve 11% total. After
this point there's little to gain with every extra attempt.