exp: add several other hashes for experimentations
crc32 has been attempted with moderate success. It might only make
sense if it's computed while reading the data so that we can reuse it.
The multiply used by lz4 was tried as well but did not provide as good
a result as the original hashes. It's not bad though. Also the existing
algorithm found by hand was slightly improved and moved to additions
instead of XOR so that it can be simplified into a multiply. Another
test was run with right shifts of 2 in each byte to get the most
interesting bits in the lower ones, but it doesn't give as good a
result.