optim: improve the distance encoding
We can save a bit shift and a conditional jump on the bits computation for
the distance code by computing both the code and the bits at once. For this
reason, the switch/case was open-coded where it's being used.

Thanks to this, we now reach 359 MB/s without CRC32 or 275 MB/s with CRC32.