optim: use a pre-computed fixed huffman distance table for most common distances
For distances below 2kB, we use a precomputed table to directly map the
distance to a code. That's easy because the encoding is fixed. Note that
the performance can further be increased by about 10% in the non-CRC32
case by mapping all 32k distances through the table. It's not yet certain
whether it could negatively impact performance on machines with small
caches, so more experimentation will be needed.

A new test shows 287 MB/s with CRC32-by-4 and 383 MB/s without CRC32.