bug: ensure we never read past the input buffer
By reading 32 bits at once, we may read past the input buffer when reading
the last bytes of the buffer. In order to fix this, we make the whole loop
stop when there are less than 4 chars left since that's the minimum we need.
And at the end of the loop, for the 1..3 remaining bytes, we jump back into
the loop where we process literals.