optim: adjust the final bytes by hand
We can simplify the code by handing the last literal bytes by hand.
We just have to increment 3 variables by the number of remaining
bytes and update the count of 9-bit codes. Better avoid jumps for
this, this helps the compiler maintain efficient code. It also
allows us to remove a test on rem >= 2 in the loop since we know
that it's >= 4 there.

A new measurement was performed with the file "daniels.html"
repeated 6800 times (500 MB), compared to last measurement :

                    Original   previous   current
    size           508891600   81806422   80219600
    time(s)             -         2.798      2.618
    Input MB/s          -           183        194
    Output MB/s         -          29.4       30.6
    BW savings MB/s     -           154        164

Both the compression ratio and the input bandwidth have increased,
resulting about 6.5% increase in BW savings.