bug/optim: don't consider the dictionary in the middle of expensive literals
If a literal string already involves a number of 9-bit codes, it will
require an expensive switch, possibly to switch back to literals. For
this reason it's preferable to avoid short matches in the middle of
complex literal sequences. The optimal situation is to consider that
if we have less than 52 9-bit codes, we know that we're still on a
fixed huffman sequence so we can switch whenever we want (hence a
match of 2 or more bytes), but if we have more 52 9-bit codes or more,
then we need at least 5 bytes to justify the switch.

It has no effect on text-based code but avoids expensive outputs in
cases where the input contains a mix of random and repetitive patterns
with binary contents, something which can happen on images for example.

In practice, it guarantees that switching between series of literals and
references cannot significantly inflate the output. A large hardly
compressible file can be produced like this :

   i=0; while ((i++<10000)); do echo -n "fo";dd if=/dev/urandom count=1 bs=7;done 2>/dev/null > a

Now with this change, the output cannot grow past the input size more
than 7 bytes per 65535 input byte, plus 16 bytes total.