optim: build an optimized CRC32 table which avoids double-negation
In crc32_char(), we're inverting the crc32 before and after computation.
These two inversions don't apply to the same part of the word, one being
used on the 24 lower bits, the other one being used with the index. This
can be optimized the following way :

	uint32_t c = ~crc;
	c = crc_table[(c & 0xff) ^ x] ^ (c >> 8);
	return ~c

   which is equivalent to :

	uint32_t c = ~crc;
	return ~(crc_table[(c & 0xff) ^ x]) ^ (~crc >> 8);

   which is equivalent to :

	uint32_t c = ~crc;
	return (crc_table[(c & 0xff) ^ x] ^ 0xFFFFFFFF) ^ (~crc >> 8);

   which is equivalent to :

	return (crc_table[(~crc & 0xff) ^ x] ^ 0xFF000000) ^ (crc >> 8);

   which is equivalent to :

	return (crc_table[(crc & 0xff ^ x) ^ 0xff] ^ 0xFF000000) ^ (crc >> 8);

   which is equivalent to :

	return (crc_table[(crc ^ x) & 0xff] ^ 0xFF000000) ^ (crc >> 8);

Thus it becomes obvious that simply swapping the index of the table and
the higher 8 bits of the word is enough to achieve this simple formula :

	return crc32_8bit[(crc ^ x) & 0xff] ^ (crc >> 8);

The net result is a performance increase of about 4%.