optim: implement a 32-bit CRC32 version working the same way as the 8-bit one
Propagating the negation is more complex here. We need 8 tables and we
need to invert 8 different bits in each table. And due to simplifications,
the computation looks bogus at first glance while it's not. The reason is
that we apply 0xff000000 to crc32_32bit[0] and since we shift it right by
8 while computing the next bytes, not inverting it effectively results in
inverting the next 8 bits.

The good point is that this method allows to use the same table for 8-bit
and for the last byte of 32-bits.

The code was implemented to support both unaligned little endian 32-bit
values and other platforms.

The result is a significant saving, about 27% faster than the byte-per-byte
mode.