fix: always include the hash part in the memmatch
We used to accept to skip the part that used to match from the hash table
but that's a bad idea as soon as we call encode() from multiple passes.
Indeed, we can find keys from a previous pass with lower offsets than the
current one and which appear valid while they are not. Thus we must always
send all input bytes through memmatch(). This fix only changes the operation
in place, but a cleaner one should be done to correctly propagate rem + 1
and mlen - 1.