fix: tell encode() whether it's encoding the last block or not
This is needed in order to let the literal encoders know what state to
switch to at the end. The BFINAL bit is now sent from slz_finish() and
not from encode() anymore, so that it's now possible to safely call
encode() as often as desired, with the only effect of reducing the
compression's efficiency.