fix: ensure that memmatch never reads past the end.
Till now there was an ambiguous case where we could make memmatch() read
past the end. This has been fixed by ensuring we never call it with less
than 4 bytes to read and by never overflowing inside it. It happens that
we always have at least 4 bytes in the caller, so the constraint is not
huge.