optim: do not always performed unaligned accesses
When processing uncompressible data, the whole data set is entirely
read at a rate of a word per input byte. This is 7% slower than doing
a per-byte read on x86_64 and 30% slower on i686. For this reason we
now make a distinction between supporting unaligned accesses and whether
or not they're faster than reading one byte and shifting it. The other
benefit is that when we resync, knowing that the CPU supports unaligned
accesses allows us to still read 32 bits at once instead of reading 3
distinct bytes.