cleanup: rearrange the main condition to avoid tests in the "if"
The "if" condition in the encode() function was becoming a bit complex
and made it difficult to apply smarter conditions to decide to encode
a reference or a literal. Let's break this. The patch looks huge but
it's just caused by the reindent caused by the removal of the first
"if" block. As a side effect, the code is slightly faster (about 1%),
probably because gcc sees better how to optimize it.