optim: avoid to encode references when they are larger than the literals
Sometimes it is possible that a reference is larger than the literal it
replaces, if it's a small literal found at a large distance. With this
commit, we estimate the encoded size of the distance and length, and
decide, compared to the length of the same block encoded as bytes, if
it's preferable to encode a reference or a literal. Note that literals
may be encoded as 8 or 9 bits. We don't want to evaluate that now, so
instead we rely on bit9 which is a good indicator : if bit9 is zero, it
is very likely that it will be for the next bytes (text). If it's set,
it's possible that it will be for the next bytes (binary) so we consider
the number of 9-bit bytes already queued.

This change allows to improve the compression ratio to about 0.5% on
hardly compressible data and has little to no effect on easily
compressible data.