optim: reduce the references table size from 32k to 8k entries
When processing highly uncompressible contents, the compression
performance drops very quickly. A ratio of 1 to 3 can be observed
under some circumstances. It happens that it's caused by the table's
size. Reducing it to 16k entries limits the loss to 50% and reducing
it further to 8k reduces the loss to about 33%, which is much more
acceptable. In parallel, the compression ratio decreases but not
that much (a few percent). However if the table goes below 8k, such
as 4k, the compression ratio drops very quickly and the performance
almost does not increase. It is very likely that the ability for the
table to remain in the L1 cache counts a lot, because at 8k entries,
that's 64kB and it might stay almost entirely in the L1 data cache
on the core i5, allowing instant lookups.