optim: pre-initialize refs for each call
The memset(refs, 0) that was in the outer loop used to result in a
significant performance gain that was wrongly attributed to a
reduction in collisions resulting in less calls to memmatch(). It
turns out that whatever is written into refs[] just before calling
encode() achieves the same performance gain, even copying itself
over itself. The real reason in fact is that cache lines are
pre-populated with known values and save memory accesses later
during the lookups. The performance gain depends a lot on the
input data set given that it depends on the number of repetitive
lookups. In practice, poorly compressible data will sweep more
refs without matching and will see a better improvement than highly
compressible ones. The non-compressible data do not see any
significant change since they're dominated by the number of calls
to memmatch().