major: make encode() take a destination buffer and return the number of bytes sent
The current implementation was OK for experimentation but now we want
to get the data in a usable format. So we pass a pointer to a "large
enough" buffer to encode(), and it copies it into the stream descriptor
so that all underlying functions update the pointer every time they write
something. This removes the need for outlen and simplifies all functions.

The copy_lit_* functions used to have a test for enough room, which is
not performed anymore since this condition is guaranteed by design and
by the calling conventions.

The change is a bit large because doing this resulted in touching all
areas. So in short, &outbuf[outlen] was moved to strm->outbuf.

Some more cleanup is possible now.