optim: reorder struct slz_stream to group commonly accessed data together
queue and qbits are used all the time, outbuf a bit less. state is used
all the time as well but misaligns the struct, so let's move it just
after the first 3. crc32 is rarely used, just like ilen which only keeps
track of the total input length to report it at the end.