bug: prevent memmatch() from finding more than 258 chars
The length control was made after memmatch() returned. The problem is
that feeding it with very repetitive data such as a file full of zeroes,
it would match a long part then re-run over that part again and again,
typically BLK/258 times, resulting in a compression 10+ times slower
than needed. This was already happening a lot on some binary files such
as JPEG images, but the extra length was not large enough to significantly
affect the compression time (eg: +/- 1%).