optim: skip the first 4 bytes in the memmatch() call
We were doing this in an early version but it was bogus because the refs
table could contain old entries from previous passes. Now that the table
is reinitialized before each pass, we definitely know that if we match a
ref, then the first 4 bytes are exactly the ones we found, so it makes
sense again to skip them during the memmatch(). Note, the code is ugly,
memmatch() should be reworked to take this into account.