optim: memmatch(): use xor between the two blocks to count the bytes
We're wasting time comparing the last 1..3 bytes twice, because we end
up in the same per-byte loop when we don't match as when we reach the
last bytes of the buffer.

Instead of comparing the two words, let's xor them. That way we can
simply observe the result to know what byte didn't match and adjust
the result accordingly. This brings approximately 2% of extra
performance.