fix: fix wrong first match of 4 zeroes
Commit 72716a2 ("optim: skip the first 4 bytes in the memmatch() call")
made it possible to match a series of 4 zeroes at the wrong place when
they were located in the first 32kB of the block being compressed. The
reason is that the refs table is pre-initialized with zeroes, meaning
the last offset is also zero and matches the condition "last < ptr".

Reseting the cached data is not enough since all values are permitted,
we need to reset the pointers. A nice solution consists in filling the
refs table with 0xFF. But doing so with memset() tends to be slow because
memset() performs some alignment checks that are not needed here. Instead,
let's write a simple reset_refs() function whose sole purpose is to reset
the refs table. It fills it with (long)-1 the fastest it can and assumes
that alignment is OK and that the size is at least a multiple of 16 longs
(which is always true here).

The end result is even faster code that remains portable and that does not
exhibit this bug anymore.

New performance looks like this for 32 kB blocks and a 8 kB refs table
(previous core i5 measures were wrong due to cpufreq) :

                  output/input size       compression speed MB/s
      file       (compression ratio)   core2/3.0 GHz   core i5/3.3 GHz
   -------------+-------------------+----------------+----------------
   index.html   |   47.90% (2.09x)  |    132 MB/s    |   188 MB/s
   daniels.html |   15.95% (6.27x)  |    310 MB/s    |   471 MB/s
   silesia.bin  |   47.02% (2.13x)  |    152 MB/s    |   220 MB/s
   silesia.zip  |   99.85% (1.00x)  |    189 MB/s    |   342 MB/s
   linux.tar    |   34.68% (2.88x)  |    161 MB/s    |   222 MB/s
   noncomp.bin  |  100.04% (1.00x)  |    216 MB/s    |   343 MB/s
   -------------+-------------------+----------------+----------------