optim: take the CRC computation out of the encode loop
There's a benefit in computing the crc prior to encoding a block, it
preloads the contents into the CPU cache and speeds up the loop later,
so in the end, the encoding loop runs at optimal speed and there's no
more need for summing non-compressible bytes one at a time.

A performance gain of 4.5% has been achieved on the silesia test with
this. The other benefit is that the encoding function is now 100%
generic and does a real deflate() regardless of the CRC algorithm, if
any.