optim: use a set of shifts to avoid the modulus
This commit avoids the expensive modulus by retrofitting the number of
bytes missed between 65521 and 65536 which is easy to count : For every
sum above 65536, the modulus is offset by (65536-65521) = 15. So for
any value, we can count the accumulated extra values by dividing the
sum by 65536 and multiplying this value by (65536-65521). That's easier
with a drawing with boxes and marbles. It gives this :

   x % 65521 = (x % 65536) + (x / 65536) * (65536 - 65521)
             = (x & 0xffff) + (x >> 16) * 15.

The performance gain is only about 1% on x86, but could be better on
machines without a fast divider/multiplier.