optim: split enqueue() between 8-bit and 16-bit usages
Most places feed enqueue() with only 8 bits at a time, despite this
we used to perform validations on the fact that the resulting qbits
was greater than 16. We can eliminate one such test for the most
common case, so now we have two functions, one for 8-bit values and
one for 16-bit values. It also makes it clearer that the limit is
indeed 16 bits. A performance increase of about 1% was achieved thanks
to this.

The resulting performance on a single core of a core i5-3320M at 3.3 GHz
(turbo) is :

                 Comp.       Compression speed MB/s
     file        ratio   rfc1950    rfc1951    rfc1952
  -------------+------+----------+----------+-----------
  index.html   | 2.09 | 199 MB/s | 227 MB/s | 190 MB/s
  daniels.html | 6.27 | 515 MB/s | 753 MB/s | 457 MB/s
  silesia.bin  | 2.13 | 230 MB/s | 266 MB/s | 217 MB/s
  silesia.zip  | 1.00 | 344 MB/s | 436 MB/s | 318 MB/s
  linux.tar    | 2.88 | 243 MB/s | 284 MB/s | 229 MB/s
  noncomp.bin  | 1.00 | 344 MB/s | 436 MB/s | 317 MB/s
  -------------+-----------------+----------+-----------

All these tests were run with write() enabled and output redirected to
/dev/null. The cost of the write() system call was measured to about
1.2 microsecond per call, and there's at least one call per round during
the tests, so high bandwidth files take a hit here. For example, the
file daniels.html without write() reaches 524 MB/s, 776 MB/s and 465 MB/s
respectively.

For reference :
  - index.html is the home page of www.haproxy.org as of 2013/12/17.
    size=76799 md5=1624df5523bb3d99bfdf3020eeb59496

  - daniels.html is http://carsten.codimi.de/gzip.yaws/daniels.html as
    used in various HTTP compression/decompression benchmarks.
    size=74837 md5=c347994bbb7d756312da379fa0aa5da3

  - silesia.zip is the Silesia corpus reference file provided here at
    http://mattmahoney.net/dc/silesia.html
    size=68182744 md5=c240c17d6805fb8a0bde763f1b94cd99

  - silesia.bin is the concatenation in alphabetic order of all the
    files found in silesia.zip
    size=211938580 md5=f92ded1d2fbc27275ba3fc3bfc69c55a

  - linux.tar is the source code of linux-4.0.0-rc1 archived using
    "git archive --format=tar v4.0.0-rc1" without prefix.
    size=589844480 md5=037eb0c5eab59feeb2965164521ac596

  - noncomp.bin was obtained using the following command :
       i=0; while ((i++<10000)); do echo -n "fo"; \
            dd if=/dev/urandom count=1 bs=7;done 2>/dev/null > a
    size=90000; md5=275fd6fe148ec3e2e5faabbcb867d25c