medium: reduce the stream state to 16 bits
It was an uint32_t which is too large for 6 values, and we need some room
for one bit. Reducing the state to 8 bits causes a small slowdown because
gcc doesn't emit optimal code. However reducing it to 16 bits actually
gives a small boost (1%). Let's stick with this for now.