medium: add support for disabling compression
The stream now has a bit "level" which is set by default to enable
compression, and which, when cleared, forces to send data as literals.
This can be used to suddenly force to disable compression, for example
when it is needed to preserve resources while a gzip stream has to be
built. The impact on performance of the extra test was measured to be
null as built.