medium: store the output format in the stream and provide generic wrappers
Storing the output format in the stream makes it possible to have generic
functions which pick the proper init/encode/finish function by just looking
at the stream format. That's much more convenient for the caller which does
not have to retain that kind of information, and the test is out of the fast
path considering that we're supposed to be calling encode() with reasonable
amounts of data.

The format is initially set by slz_init() now. The direct functions continue
to work as they also set the format themselves.