fix: fix again wrong match of 4 zeroes
Commit fd44233 ("fix: fix wrong first match of 4 zeroes") was used to
fix commit 72716a2 ("optim: skip the first 4 bytes in the memmatch()
call") which caused series of 4 zeroes to accidently match in the refs
cache. Unfortunately the fix was not enough because by initializing
the refs with -1, the condition (ptr - last - 1 >= 32768) can still
accidently match for ptr <= 32768.

In order to prevent this accidental match, instead it is required to
initialize any entry to -32769, that will ensure that any positive
value from 0 to (long)-32769 cannot cause an undesired match.