bug: init() must end in EOB mode for rfc1951 (raw deflate)
The raw deflate mode doesn't send a header, and because of this the
stream's state remained in SLZ_ST_INIT state. This caused it to start
with an extra EOB when sending literals, or conversely to miss the
start bits with BTYPE/BFINAL before the first ref was found.

Given how small the init functions are, it's best to split them and
to perform the format-specific assignments in each of them. Anyway
it's already partially the case (crc and format).