medium: send the gzip/zlib header before sending data or finishing
This patch ensures that the gzip/zlib header is always sent before
sending the first piece of data. Till now this was handled by the
init function but the purpose is to ensure the init function does
not need a reference to the buffer, so it will not have to do it
anymore soon.